spring-boot-add-category2

+++ src/main/java/mate/academy/bookstore/service/CategoryServiceImpl.java
@@ -45,6 +45,8 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void deleteById(Long id) {
-        categoryRepository.deleteById(id);
+        if (categoryRepository.findById(id).isPresent()) {
+            categoryRepository.deleteById(id);
+        }
     }
 }
